requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
lxml==4.6.3
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
spyne==2.13.16
sqlparse==0.4.1
typing-extensions==3.10.0.0
